fix: make events annotation an optional plugin
